Reorganized the cogs and its commands
- Now there is a new cog for info related stuff.
