+++ cogs/utilities.py
@@ -2,107 +2,7 @@ import datetime
 import discord
 import emoji
 from discord.ext import commands
-from discord.ext import menus
-from discord import ui
-import requests
 import asyncio
-
-class MyMenuPages(ui.View, menus.MenuPages):
-    def __init__(self, source, page):
-        super().__init__(timeout=60)
-        self._source = source
-        self.current_page = page
-        self.ctx = None
-        self.message = None
-
-    async def start(self, ctx, *, channel=None, wait=False):
-        # We wont be using wait/channel, you can implement them yourself. This is to match the MenuPages signature.
-        await self._source._prepare_once()
-        self.ctx = ctx
-        self.message = await self.send_initial_message(ctx, ctx.channel)
-
-    async def _get_kwargs_from_page(self, page):
-        """This method calls ListPageSource.format_page class"""
-        value = await super()._get_kwargs_from_page(page)
-        if 'view' not in value:
-            value.update({'view': self})
-        return value
-
-    async def interaction_check(self, interaction):
-        """Only allow the author that invoke the command to be able to use the interaction"""
-        return interaction.user == self.ctx.author
-
-    # This is extremely similar to Custom MenuPages(I will not explain these)
-    @ui.button(emoji='⏪', style=discord.ButtonStyle.blurple)
-    async def first_page(self, interaction, button):
-        await self.show_page(0)
-        await interaction.response.defer()
-
-    @ui.button(emoji='◀️', style=discord.ButtonStyle.blurple)
-    async def before_page(self, interaction, button):
-        await self.show_checked_page(self.current_page - 1)
-        await interaction.response.defer()
-
-    @ui.button(emoji='⏹', style=discord.ButtonStyle.blurple)
-    async def stop_page(self, interaction, button):
-        self.stop()
-        await interaction.response.defer()
-
-    @ui.button(emoji='▶️', style=discord.ButtonStyle.blurple)
-    async def next_page(self, interaction, button):
-        await self.show_checked_page(self.current_page + 1)
-        await interaction.response.defer()
-
-    @ui.button(emoji='⏩', style=discord.ButtonStyle.blurple)
-    async def last_page(self, interaction, button):
-        await self.show_page(self._source.get_max_pages() - 1)
-        await interaction.response.defer()
-
-
-
-class MySource(menus.ListPageSource):
-    def __init__(self, ctx, entries, per_page):
-        self.ctx = ctx
-        self.per_page = per_page
-        self.entries = entries
-        pages, left_over = divmod(len(entries), per_page)
-        if left_over:
-            pages += 1
-
-        self._max_pages = pages
-
-    async def format_page(self, menu, entries):
-        offset = (menu.current_page * self.per_page) + 1
-        bababoi= [list(x) for x in enumerate(entries, start=offset)]
-        
-        for x in bababoi[:3]:
-            if x[0] == 1:
-                x[0] = '🥇'
-            elif x[0] == 2:
-                x[0] = '🥈'
-            elif x[0] == 3:
-                x[0] = '🥉'
-
-        total_data = len(self.entries)
-        total = f"{offset:,} of {min(total_data, offset + self.per_page -1):,} of {total_data:,} banned users"
-
-        if menu.current_page==0:
-            first_list = ["{0} - {1} - {2}".format(x[0],x[1][0],x[1][1]) for x in bababoi[:3]]
-            second_list = ["#{0} - {1} - {2}".format(x[0],x[1][0],x[1][1]) for x in bababoi[3:]]
-            first_list.extend(second_list)
-
-        else:
-            first_list = ["#{0} - {1} - {2}".format(x[0],x[1][0],x[1][1]) for x in bababoi]
-
-
-        e = discord.Embed(
-            title=f"👷‍♂️ Classifica Builder",
-            description="\n".join(first_list),
-            color=discord.Color.yellow()
-            )
-
-        e.set_footer(text=f"Page {(menu.current_page) + 1}/{(self._max_pages)} · Top 10 Builer BTE Italia ")
-        return e
 
 
 class Utilities(commands.Cog):
@@ -453,76 +353,6 @@ class Utilities(commands.Cog):
                 color=discord.Color.blue()
                 )
             await ctx.send(embed=embed)
-
-
-
-    @commands.command(
-        name='progressi',
-        description='Will send link to BTE Italias progress map.',
-        usage='£progressi',
-        brief='BTE Italia progress map',
-        aliases=["progress", "map", "mappa"]
-    )
-    async def progressi(self, ctx):         
-            embed = discord.Embed(
-                description=":map: **Ecco la mappa progressi!**\nhttps://maphub.net/BTEItalia/bte-italia-progressi", color=discord.Color.blue())
-            await ctx.send(embed=embed)     
-
-
-
-    @commands.command(
-        name='staff',
-        description='Will send a list of all BTE Italias staff member.',
-        usage='£progressi',
-        brief='BTE Italia staff member list',
-        aliases=["stafflist", "mods", "moderatori", "supporto", "modlist", "lista mod", "lista moderatori", "tecnici", "lista staff", "list", "staffs", "staffer", "staffers"]
-    )
-    async def staff(self, ctx):         
-        message = discord.Embed(title="<:bte_italy:991738968725000433> Lista Staff BTE Italia ", colour=discord.Colour.blue())
-        message.add_field(name="<:presidente:1010588696573128784> Presidente", value="<:magix92:1010586235535888434> `Magix92`\n<:filippo_the_king:1010586249817497753> `filippo_the_king`", inline=False)
-        message.add_field(name="<:teamlead:1010588694832500757> Team Lead", value="<:forgiobombi:1010585405889974312> `forgiobombi`\n<:Diamondpower500:1010587025205907516> `DiamondPower500`\n<:Martinsuperstar:1010590112205250580> `Martinsuperstar`\n<:MemoryOfLife:1010590510760611930> `MemoryOfLife`", inline=False)
-        message.add_field(name="<:tecnico:1010588693423194233> Tecnico", value="<:forgiobombi:1010585405889974312> `forgiobombi`\n<:MemoryOfLife:1010590510760611930> `MemoryOfLife`\n<:filippo_the_king:1010586249817497753> `filippo_the_king`\n<:ElijahPepe:1010592808836546572> `ElijahPepe`\n<:MikChan:1010592810380054658> `MikChan`\n<:celery6:1010593496526229525> `celery6`\n<:Marcy2005:1010597530884636706> `spl1ce_`", inline=False)
-        message.add_field(name="<:modesupporto:1010588684875202680> Mod e Supporto", value="<:Diamondpower500:1010587025205907516> `DiamondPower500`\n<:I_EricDraven_I:1010595264492470314> `I_EricDraven_I`\n<:FedexV2:1010595261535498330> `FedeXV2`\n<:HannibalLecter:1010595263401955459> `HannibalLecter`", inline=False)
-        message.add_field(name="<:valutazioni:1010588685831524353> Valutazioni", value="<:Martinsuperstar:1010590112205250580> `Martinsuperstar`\n<:forgiobombi:1010585405889974312> `forgiobombi`\n<:magix92:1010586235535888434> `Magix92`", inline=False)
-        message.add_field(name="<:personalerelazionipubbliche:1010588691388977153> Personale Relazioni Pubbliche", value="<:forgiobombi:1010585405889974312> `forgiobombi`\n<:I_EricDraven_I:1010595264492470314> `I_EricDraven_I`\n<:magix92:1010586235535888434> `Magix92`\n<:H2bomber_:1010603700412219523> `H2bomber_`\n<:howard9068:1010603701817331844> `howard9068`\n<:glurbB:1010603698730315787> `glurbB`", inline=False)
-        await ctx.send(embed=message)  
-        
-
-
-    @commands.command(
-        name='ip',
-        description='Will send the ip to the BTE Italia Minecraft server.',
-        usage='£ip',
-        brief='BTE Italia Minecraft server IP',
-        aliases=["adress", "inzirizzo", "server", "ipmc", "ipserver"]
-    )
-    async def ip(self, ctx):         
-            embed = discord.Embed(
-                description="<:bte_italy:991738968725000433> **Ecco gli IP per il nostro Server Minecraft!**\n\n<:java:1010963036342861824> **Java**: `mc.bteitalia.tk`\n<:bedrock:1010963327654055937> **Bedrock**: `bedrock.buildtheearth.net`", color=discord.Color.blue())
-            await ctx.send(embed=embed)     
-
-
-
-    @commands.command(
-        name='minecraft_leaderboard',
-        aliases=['mclb', 'mc_lb']
-    )
-    async def minecraft_leaderboard(self, ctx, page=1):
-        
-        r = requests.get("https://62.171.174.31:8000/points", verify=False)
-        data = r.json()
-        lb_data = data['Leaderboard']
-        sorted_lb_data = sorted(lb_data, key=lambda d: d['score'],  reverse=True)
-        
-        lb_list=[]
-        for x in sorted_lb_data:
-            values = list(x.values())
-            lb_list.append(values)
-
-
-        formatter = MySource(ctx, lb_list, per_page=10)
-        menu = MyMenuPages(formatter, page-1)
-        await menu.start(ctx)
